// src/types/Basic.d.ts
// 基本的な音符の型
export type Note = "C" | "C#" | "D" | "D#" | "E" | "F" | "F#" | "G" | "G#" | "A" | "A#" | "B";

// ディグリー
export type Degree = "Ⅰ" | "Ⅱ" | "Ⅲ" | "Ⅳ" | "Ⅴ" | "Ⅵ" | "Ⅶ";

// コードの基本タイプ
export type ChordType = "major" | "minor" | "dim" | "aug";

export type Extension = "6" | "7" | "maj7";

export type AddNote = "add9" | "add11" | "add13";